requirements-audit.txt
pytest
pytest-django
pytest-xdist
bandit
safety
requests
colorama
